Add Mat-Dialoge when delete Product

+++ src/app/products/Components/product-item/product-item.component.ts
@@ -1,5 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { ProductService } from '../../Services/product.service';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogeComponent } from '../../../material/dialoge/dialoge.component';
 
 @Component({
   selector: 'app-product-item',
@@ -9,11 +11,25 @@ import { ProductService } from '../../Services/product.service';
 export class ProductItemComponent implements OnInit {
   @Input() prd!: any;
 
-  constructor(private productService:ProductService) { }
+  constructor(private productService:ProductService,
+              private dialog: MatDialog) { }
 
   ngOnInit(): void {
   }
   deleteProduct(code:string){
-    this.productService.deleteProduct(code);
+    const dialogRef = this.dialog.open(DialogeComponent, {
+      width: '400px',
+      height:'220px',
+      data: { message: this.prd.name },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result === 'confirm') {
+
+        this.productService.deleteProduct(code);
+        window.location.reload();
+          }
+    });
   }
+
 }
